pyproject.toml
[tool.poetry]
name = "fusefs"
version = "0.1.0"
description = ""
authors = ["Miroslav Bauer <bauer@cesnet.cz>"]

[tool.poetry.scripts]
fusefs = 'fusefs:main'

[tool.poetry.dependencies]
python = "^3.7"
fs = "^2.2.1"
fusepy = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
